# adidas-db/api-docs.json
{"swagger":"2.0","info":{"description":"Microservice to acess db","version":"1.0","title":"AdidasDB","license":{"name":"Iorran License"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"subscription-controller","description":"Subscription Controller"}],"paths":{"/subscriptions":{"get":{"tags":["subscription-controller"],"summary":"Recovery all subscriptions","operationId":"findAllSubscription","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Subscription"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"basicAuth":[]}],"deprecated":false},"post":{"tags":["subscription-controller"],"summary":"Create a new subscription","operationId":"createSubscription","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"subscription","description":"subscription","required":true,"schema":{"$ref":"#/definitions/Subscription"}}],"responses":{"200":{"description":"OK","schema":{"type":"integer","format":"int64"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"basicAuth":[]}],"deprecated":false}},"/subscriptions/{id}":{"get":{"tags":["subscription-controller"],"summary":"Recovery subscription by id","operationId":"findSubscriptionById","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Subscription"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"basicAuth":[]}],"deprecated":false}}},"securityDefinitions":{"basicAuth":{"type":"basic"}},"definitions":{"Subscription":{"type":"object","required":["consent","dateOfBith","email","newsletterId"],"properties":{"consent":{"type":"boolean","description":"flag for consent"},"dateOfBith":{"type":"string","format":"date","description":"user date of bith"},"email":{"type":"string","description":"user email"},"firstName":{"type":"string","description":"user name"},"gender":{"type":"string","description":"gender","enum":["M","F"]},"id":{"type":"integer","format":"int64","description":"subscription id generated value"},"newsletterId":{"type":"integer","format":"int64","description":"newsletter Id corresponding to the campaign"}},"title":"Subscription"}}}